backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10
celery[redis]==5.4.0
redis==5.2.1
httpx==0.28.1
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
oauth2client==4.1.3
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
python-dotenv==1.0.1
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
